Now using Cairo for drawing of potential energy surface.
Cairo gives us (1) finer-grained control over the rendering (e.g. ability to
determine width/height of text) and (2) ability to output to multiple contexts
(e.g. PDF or SVG files). Cairo is free and open source. (Incidentally, it is
what OpenBabel uses to draw molecules, so if you've installed that, then
you're all set for this as well.